requirements.txt
Django==1.8.5
django-haystack==2.4.1
wheel==0.24.0
Whoosh==2.7.4
